rm bloat, add single grass js file
